requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
pydantic==2.4.2
pydantic-settings==2.0.3
pydantic[email]==2.4.2
twilio==8.10.0
fastapi-mail==1.4.1
python-multipart==0.0.6
email-validator==2.1.0